lightningd/onchain_control: routines to control onchaind (move from peer_control.c)
