+++ lightningd/Makefile
@@ -63,6 +63,7 @@ LIGHTNINGD_SRC :=				\
 	lightningd/log.c			\
 	lightningd/log_status.c			\
 	lightningd/netaddress.c			\
+	lightningd/onchain_control.c		\
 	lightningd/opening_control.c		\
 	lightningd/opt_time.c			\
 	lightningd/options.c			\
